segmented scoresheet page
