+++ package.json
@@ -9,6 +9,7 @@
     "react-flexbox-grid": "^2.0.0",
     "react-redux": "^5.0.6",
     "react-redux-firebase": "^2.0.0-rc.2",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
     "redux": "^3.7.2"
   },
